Add top-level Git class
