+++ lib/project_types/node/commands/create.rb
@@ -81,7 +81,7 @@ module Node
       end
 
       def build(name)
-        ShopifyCli::Helpers::Git.clone('https://github.com/Shopify/shopify-app-node.git', name)
+        ShopifyCli::Git.clone('https://github.com/Shopify/shopify-app-node.git', name)
         ShopifyCli::Finalize.request_cd(name)
 
         @ctx.root = File.join(@ctx.root, name)
