+++ lib/shopify-cli/helpers.rb
@@ -5,7 +5,6 @@ module ShopifyCli
     autoload :Async, 'shopify-cli/helpers/async'
     autoload :EnvFile, 'shopify-cli/helpers/env_file'
     autoload :FS, 'shopify-cli/helpers/fs'
-    autoload :Git, 'shopify-cli/helpers/git'
     autoload :GraphQL, 'shopify-cli/helpers/graphql'
     autoload :Haikunator, 'shopify-cli/helpers/haikunator'
     autoload :Node, 'shopify-cli/helpers/node'
