+++ test/shopify-cli/api_test.rb
@@ -19,7 +19,7 @@ module ShopifyCli
         token: 'faketoken',
         url: "https://my-test-shop.myshopify.com/admin/api/2019-04/graphql.json",
       )
-      Helpers::Git.stubs(:sha).returns('abcde')
+      Git.stubs(:sha).returns('abcde')
       @api.stubs(:uname).with(flag: 'v').returns('Mac')
     end
